backend/requirements.txt
fastapi
uvicorn
pandas
google-generativeai
joblib
scikit-learn==1.7.2
tabulate
numpy==2.4.3
python-dotenv
openai
httpx==0.27.2
PyYAML==6.0.1
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-experimental>=0.3.0
langchain-openai>=0.2.0